refactor(ramps): move fiat deposit payment-method helpers into the Ramp tree
useFiatDepositPaymentMethods and fiatDepositAsset are ramps logic that MM Pay
consumes, so they belong under app/components/UI/Ramp with money-movement
ownership rather than in the confirmations tree. Import paths only; no
behaviour change.
