+++ app/components/UI/Ramp/utils/fiatDepositAsset.test.ts
