+++ app/components/Views/confirmations/hooks/pay/useAutomaticTransactionPayToken.ts
@@ -44,7 +44,7 @@ import { selectPaymentOverrideByTransactionId } from '../../../../../selectors/t
 import { MUSD_TOKEN_ADDRESS } from '../../../../UI/Earn/constants/musd';
 import { useWithdrawTokenFilter } from './useWithdrawTokenFilter';
 import { useTransactionAccountOverride } from '../transactions/useTransactionAccountOverride';
-import { useFiatDepositPaymentMethods } from './useFiatDepositPaymentMethods';
+import { useFiatDepositPaymentMethods } from '../../../../UI/Ramp/hooks/useFiatDepositPaymentMethods';
 
 export interface SetPayTokenRequest {
   address: Hex;
