+++ app/components/Views/confirmations/hooks/pay/useTransactionPaySelectedFiatPaymentMethod.test.ts
@@ -2,10 +2,10 @@ import { renderHook } from '@testing-library/react-hooks';
 import { type PaymentMethod } from '@metamask/ramps-controller';
 import { useTransactionPaySelectedFiatPaymentMethod } from './useTransactionPaySelectedFiatPaymentMethod';
 import { useTransactionPayFiatPayment } from './useTransactionPayData';
-import { useFiatDepositPaymentMethods } from './useFiatDepositPaymentMethods';
+import { useFiatDepositPaymentMethods } from '../../../../UI/Ramp/hooks/useFiatDepositPaymentMethods';
 
 jest.mock('./useTransactionPayData');
-jest.mock('./useFiatDepositPaymentMethods');
+jest.mock('../../../../UI/Ramp/hooks/useFiatDepositPaymentMethods');
 
 const PAYMENT_METHOD_MOCK = {
   id: 'pm-123',
